Added Raw fencing data to the repository with VAP
Needed to have more detailed naming for shapefile creation. changed the processed files to explicitly note the Herd and the fence.

Added code to potentially use rgeos instead of sf for buffering. may require full rewrite if I cant convert the sf object to sp.

Change the characters used to deliniate between herd and fence in shapefile name to use _ instead of ,.